- moved gpu/testutil docker related helpers into /pkg/util/testutil/docker - refactored util/testutil/docker pkg to have run.go and get.go files
